authorisation

+++ routes/campground.js
@@ -2,23 +2,9 @@ const express = require('express');
 const router = express.Router();
 
 const catchAsync = require('../utils/catchAsync');
-const ExpressError = require('../utils/ExpressError');
-const { campgroundSchema, reviewSchema } = require('../schemas.js');
-const isLoggedIn = require('../middleware');
+const { isLoggedIn, isAuthor, validateCampground } = require('../middleware');
 
 const Campground = require('../models/campground');
-
-
-const validateCampground = (req, res, next) => {
-    const { error } = campgroundSchema.validate(req.body);
-    if (error){
-        const msg = error.details.map(el => el.message).join(',');
-        throw new ExpressError(msg, 400)
-    } else {
-        next();
-    };
-};
-
 
 
 router.get('/', catchAsync( async (req, res) => {
@@ -32,38 +18,58 @@ router.get('/new', isLoggedIn, (req, res) => {
 
 router.post('/', isLoggedIn, validateCampground, catchAsync( async (req, res, next) => {
     const campground = new Campground(req.body.campground);
+    campground.author = req.user._id;
     await campground.save();
     req.flash('success', 'Succesfully made a new Campground!');
     res.redirect(`/campgrounds/${ campground._id}`);
 }));
 
 router.get('/:id', catchAsync( async (req, res) => {
-    const campground = await Campground.findById(req.params.id).populate('reviews');
-        if (!campground) {
-            req.flash('error', 'Campground not found');
-            return res.redirect('/campgrounds');
-        };
-    res.render('campgrounds/show', { campground, title: campground.title });
-}));
-
-router.get('/:id/edit', isLoggedIn, catchAsync( async (req, res) => {
-    const campground = await Campground.findById(req.params.id);
-    res.render('campgrounds/edit', { campground, title: 'Edit Campground' });
-}));
-
-router.put('/:id', isLoggedIn, validateCampground, catchAsync( async (req, res) => {
-    const { id } = req.params;
-    const campground = await Campground.findByIdAndUpdate(id, {...req.body.campground});
+    const campground = await (await Campground.findById(req.params.id).populate({
+        path: 'reviews',
+        populate: {
+            path: 'author'
+        }
+        }).populate('author'));
     if (!campground) {
         req.flash('error', 'Campground not found');
         return res.redirect('/campgrounds');
     };
+    res.render('campgrounds/show', { campground, title: campground.title });
+}));
+
+router.get('/:id/edit', isLoggedIn, isAuthor, catchAsync( async (req, res) => {
+    const campground = await Campground.findById(req.params.id);
+    if (!campground) {
+        req.flash('error', 'Campground not found');
+        return res.redirect('/campgrounds');
+    };
+    res.render('campgrounds/edit', { campground, title: 'Edit Campground' });
+}));
+
+router.put('/:id', isAuthor, isLoggedIn, validateCampground, catchAsync( async (req, res) => {
+    const { id } = req.params;
+    const campground = await Campground.findById(id);
+    if (!campground) {
+        req.flash('error', 'Campground not found');
+        return res.redirect('/campgrounds');
+    };
+    const camp = await Campground.findByIdAndUpdate(id, {...req.body.campground});
     req.flash('success', 'Successfully updated Campground');
     res.redirect(`/campgrounds/${ campground._id }`);
 }));
 
 router.delete('/:id', isLoggedIn, catchAsync( async (req, res) => {
     const { id } = req.params;
+    const campground = await Campground.findById(id);
+    if (!campground) {
+        req.flash('error', 'Campground not found');
+        return res.redirect('/campgrounds');
+    };
+    if (!campground.author.equals(req.user._id)) {
+        req.flash('error', 'You are not authorized to do this');
+        return res.redirect(`/campgrounds/${ campground._id }`);
+    };
     await Campground.findByIdAndDelete(id);
     req.flash('success', 'Successfully deleted the Campground');
     res.redirect('/campgrounds');
